logger: move to helpers package
So we can use it in other packages, not just "frontend"

At this point, I didn't think it was worth it to make a new "logger"
package, so I moved it under helpers.
