+++ frontend/lib/swr-cache/index.ts
@@ -16,7 +16,7 @@
  */
 
 import { APP_ORIGIN } from '@config/env';
-import { log as defaultLog, Logger, nullLog } from '@lib/logger';
+import { log as defaultLog, Logger, nullLog } from '@ifixit/helpers/logger';
 import type { NextApiHandler } from 'next';
 import { z } from 'zod';
 import { getCache } from './adapters';
